refactor(capabilities): :recycle: dissolve design-system/utils — JsonLd to features/seo, mermaid-loader colocated, shuffleArray to lib
