+++ src/components/content/blog/blog-generic-post-list-page-template/blog-generic-post-list-page-template.tsx
@@ -4,7 +4,7 @@ import { Content } from "@/types/content/content";
 import { tracking } from "@/types/configuration/tracking";
 import { FC } from "react";
 import { ContentPage } from "@/components/features/content/content-page";
-import { JsonLd } from "@/components/design-system/utils/components/jsond-ld";
+import { JsonLd } from "@/components/features/seo/jsond-ld";
 import { siteMetadata } from "@/types/configuration/site-metadata";
 
 export interface BlogGenericPostListPageProps {
