+++ src/components/content/data-structures-and-algorithms/roadmap/roadmap.tsx
@@ -5,7 +5,7 @@ import { ReadingContentPage } from "@/components/features/content/reading-conten
 import { siteMetadata } from "@/types/configuration/site-metadata";
 import { tracking } from "@/types/configuration/tracking";
 import { FC, PropsWithChildren } from "react";
-import { JsonLd } from "@/components/design-system/utils/components/jsond-ld";
+import { JsonLd } from "@/components/features/seo/jsond-ld";
 import { Content } from "@/types/content/content";
 
 interface DsaProps {
