+++ src/components/content/data-structures-and-algorithms/topic/topic.tsx
@@ -8,7 +8,7 @@ import { slugs } from "@/types/configuration/slug";
 import { tracking } from "@/types/configuration/tracking";
 import { FC, PropsWithChildren } from "react";
 import { CourseNavigation } from "@/components/content/data-structures-and-algorithms/course-navigation";
-import { JsonLd } from "@/components/design-system/utils/components/jsond-ld";
+import { JsonLd } from "@/components/features/seo/jsond-ld";
 import { Content } from "@/types/content/content";
 
 interface DsaProps {
