+++ src/components/content/home/homepage/homepage.tsx
@@ -1,6 +1,6 @@
 import { MatrixBackground } from "@/components/design-system/molecules/effects/matrix-background";
 import { Menu } from "@/components/design-system/organism/menu";
-import { JsonLd } from "@/components/design-system/utils/components/jsond-ld";
+import { JsonLd } from "@/components/features/seo/jsond-ld";
 import { ProfilePresentation } from "./profile-presentation";
 import { siteMetadata } from "@/types/configuration/site-metadata";
 import { tracking } from "@/types/configuration/tracking";
