+++ src/components/design-system/molecules/diagram/mermaid-diagram/mermaid-loader.ts
