--- src/components/design-system/utils/shuffle-array.ts
@@ -1,4 +1,0 @@
-export const shuffleArray = <T>(array: T[] = [], numberOfItems: number) =>
-    [...array]
-          .sort(() => Math.random() - 0.5)
-          .slice(0, numberOfItems)
