+++ src/components/features/seo/jsond-ld/index.ts
@@ -1,0 +1,1 @@
+export { JsonLd } from "./jsond-ld";
