+++ src/lib/shuffle-array/shuffle-array.ts
@@ -1,0 +1,2 @@
+export const shuffleArray = <T>(array: T[] = [], numberOfItems: number) =>
+    [...array].sort(() => Math.random() - 0.5).slice(0, numberOfItems);
